extract graphql business logic from send feedback and put it in graphql service
